refactor: group API
