+++ src/api/friend-group/service/friend-group.service.ts
@@ -1,8 +1,4 @@
-import {
-  BadRequestException,
-  Injectable,
-  InternalServerErrorException,
-} from '@nestjs/common';
+import { Injectable, InternalServerErrorException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { FriendGroupEntity } from '../model/friend-group.entity';
 import { Repository } from 'typeorm';
@@ -10,6 +6,7 @@ import { FriendGroupDto } from '../model/friend-group.dto';
 import { FriendGroupResponse } from '../model/friend-group.response';
 import { FriendEntity } from '../../friend/model/friend.entity';
 import { FriendGroupSaveResponse } from '../model/friend-group-save.response';
+import { FriendGroupHelpService } from './friend-group-help-service';
 
 @Injectable()
 export class FriendGroupService {
@@ -18,38 +15,16 @@ export class FriendGroupService {
     private friendGroupEntityRepository: Repository<FriendGroupEntity>,
     @InjectRepository(FriendEntity)
     private friendListEntityRepository: Repository<FriendEntity>,
+    private readonly friendGroupHelpService: FriendGroupHelpService,
   ) {}
 
   async findAllGroupsBy(memberId: number): Promise<FriendGroupResponse[]> {
     const groups = await this.friendGroupEntityRepository.findBy({ memberId });
-    return Promise.all(
-      groups.map(async (group) => {
-        const groupMemberCount = await this.friendListEntityRepository.countBy({
-          groupId: group.id,
-        });
-        const friendsInGroup = await this.friendListEntityRepository.findBy({
-          groupId: group.id,
-        });
-        const thumbnailImageUrls = this.getThumbnailImageUrls(friendsInGroup);
-
-        return {
-          groupId: group.id,
-          groupName: group.name,
-          groupMemberCount,
-          thumbnailImageUrls,
-        };
-      }),
-    );
-  }
-
-  private getThumbnailImageUrls(friendsInGroup: FriendEntity[]) {
-    return friendsInGroup
-      .map((friend) => friend.thumbnailImageUrl)
-      .filter((url): url is string => !!url);
+    return Promise.all(this.friendGroupHelpService.getGroupDetails(groups));
   }
 
   async createGroup(dto: FriendGroupDto): Promise<FriendGroupSaveResponse> {
-    await this.assertDuplicatedGroup(dto);
+    await this.friendGroupHelpService.assertDuplicatedGroup(dto);
     const { id } = await this.friendGroupEntityRepository.save(dto);
     if (!id) {
       throw new InternalServerErrorException('저장에 실패했습니다.');
@@ -73,18 +48,6 @@ export class FriendGroupService {
       };
     } catch (e) {
       throw new InternalServerErrorException('그룹명 변경에 실패했습니다.');
-    }
-  }
-
-  private async assertDuplicatedGroup({ memberId, name }: FriendGroupDto) {
-    const existedGroups = await this.friendGroupEntityRepository.findBy({
-      memberId,
-    });
-    const duplicatedGroup = existedGroups.find((group) => group.name === name);
-    if (duplicatedGroup) {
-      throw new BadRequestException(
-        '이미 존재하는 그룹명을 사용할 수 없습니다.',
-      );
     }
   }
 }
